requirements.txt
flask==3.0.0
edge-tts==6.1.9
Pillow==10.2.0
requests==2.31.0
